// cypress/e2e/bms/common/types.ts
export type TUserRole =
  | "BM"
  | "SBM"
  | "VP"
  | "VP OPS"
  | "TGA OPS"
  | "Finance"
  | "Viewer";

export type TChannel = "marketplace" | "offline";

export type TDeliveryMethod = "Gudang Ada Logistic" | "Dikirim Penjual";

export type TSettingType = "margin" | "price";

export type TTaxType = "Non PKP" | "PKP";